Updated structure. Moved XML code to xml folder in lib
