+++ lib/sweet-data.js
@@ -34,72 +34,7 @@ pp.prototype.setStep = function(steps, stepChar){
 	
 // ----------------------- XML section ----------------------------------------------------
 
-pp.prototype.xml = function(text) {
-
-	var ar = text.replace(/>\s{0,}</g,"><")
-				 .replace(/</g,"~::~<")
-				 .replace(/xmlns\:/g,"~::~xmlns:")
-				 .replace(/xmlns\=/g,"~::~xmlns=")
-				 .split('~::~'),
-		len = ar.length,
-		inComment = false,
-		deep = 0,
-		str = '',
-		ix = 0;
-
-		for(ix=0;ix<len;ix++) {
-			// start comment or <![CDATA[...]]> or <!DOCTYPE //
-			if(ar[ix].search(/<!/) > -1) { 
-				str += this.shift[deep]+ar[ix];
-				inComment = true; 
-				// end comment  or <![CDATA[...]]> //
-				if(ar[ix].search(/-->/) > -1 || ar[ix].search(/\]>/) > -1 || ar[ix].search(/!DOCTYPE/) > -1 ) { 
-					inComment = false; 
-				}
-			} else 
-			// end comment  or <![CDATA[...]]> //
-			if(ar[ix].search(/-->/) > -1 || ar[ix].search(/\]>/) > -1) { 
-				str += ar[ix];
-				inComment = false; 
-			} else 
-			// <elm></elm> //
-			if( /^<\w/.exec(ar[ix-1]) && /^<\/\w/.exec(ar[ix]) &&
-				/^<[\w:\-\.\,]+/.exec(ar[ix-1]) == /^<\/[\w:\-\.\,]+/.exec(ar[ix])[0].replace('/','')) { 
-				str += ar[ix];
-				if(!inComment) deep--;
-			} else
-			 // <elm> //
-			if(ar[ix].search(/<\w/) > -1 && ar[ix].search(/<\//) == -1 && ar[ix].search(/\/>/) == -1 ) {
-				str = !inComment ? str += this.shift[deep++]+ar[ix] : str += ar[ix];
-			} else 
-			 // <elm>...</elm> //
-			if(ar[ix].search(/<\w/) > -1 && ar[ix].search(/<\//) > -1) {
-				str = !inComment ? str += this.shift[deep]+ar[ix] : str += ar[ix];
-			} else 
-			// </elm> //
-			if(ar[ix].search(/<\//) > -1) { 
-				str = !inComment ? str += this.shift[--deep]+ar[ix] : str += ar[ix];
-			} else 
-			// <elm/> //
-			if(ar[ix].search(/\/>/) > -1 ) { 
-				str = !inComment ? str += this.shift[deep]+ar[ix] : str += ar[ix];
-			} else 
-			// <? xml ... ?> //
-			if(ar[ix].search(/<\?/) > -1) { 
-				str += this.shift[deep]+ar[ix];
-			} else 
-			// xmlns //
-			if( ar[ix].search(/xmlns\:/) > -1  || ar[ix].search(/xmlns\=/) > -1) { 
-				str += this.shift[deep]+ar[ix];
-			} 
-			
-			else {
-				str += ar[ix];
-			}
-		}
-		
-	return  (str[0] == '\n') ? str.slice(1) : str;
-}
+pp.prototype.xml = require('./xml/index').xml;
 
 // ----------------------- JSON section ----------------------------------------------------
 
@@ -259,12 +194,7 @@ pp.prototype.sql = function(text) {
 
 // ----------------------- min section ----------------------------------------------------
 
-pp.prototype.xmlmin = function(text, preserveComments) {
-
-	var str = preserveComments ? text
-				   : text.replace(/\<![ \r\n\t]*(--([^\-]|[\r\n]|-[^\-])*--[ \r\n\t]*)\>/g,"");
-	return  str.replace(/>\s{0,}</g,"><"); 
-}
+pp.prototype.xmlmin = require('./xml/index').xmlmin;
 
 pp.prototype.jsonmin = function(text) {
